# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Open Civic Signal OS API
  version: 0.2.0
paths:
...
  /api/signals/prioritized:
    get:
      summary: Get prioritized civic backlog
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          required: false
          description: Optional comma-separated lifecycle status filter (for example `NEW` or `NEW,IN_PROGRESS`).
          schema:
            type: string
            example: NEW
        - name: X-Community-Id
          in: header
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '200':
          description: Paginated prioritized list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPrioritizedSignals'
  /api/signals/top-10:
    get:
      summary: Get top 10 unresolved civic problems
      parameters:
        - name: X-Community-Id
          in: header
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '200':
          description: List of top signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrioritizedSignal'
  /api/signals/{id}:
    get:
      summary: Get one civic signal detail by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Community-Id
          in: header
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Signal detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalResponse'
        '400':
          description: Invalid signal ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Signal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/signals/meta:
    get:
      summary: Get aggregate metadata for current signal scope
      parameters:
        - name: X-Community-Id
          in: header
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Aggregated signal metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalMeta'
  /api/signals/mine:
    get:
      summary: Get signals created by the authenticated user
      parameters:
        - name: X-Community-Id
          in: header
          required: false
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated self-reported signals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPrioritizedSignals'
  /api/signals/{id}/vote:
    post:
      summary: Cast support vote for a signal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Community-Id
          in: header
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Updated signal after vote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalResponse'
        '404':
          description: Signal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
...
    ScoreBreakdown:
      type: object
      required:
        - urgency
        - impact
        - affectedPeople
        - communityVotes
      properties:
        urgency:
          type: number
        impact:
          type: number
        affectedPeople:
          type: number
        communityVotes:
          type: number
    PrioritizedSignal:
      type: object
      required:
        - id
        - title
        - category
        - scoreBreakdown
        - priorityScore
        - status
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        scoreBreakdown:
          $ref: '#/components/schemas/ScoreBreakdown'
        priorityScore:
          type: number
        status:
          type: string
          enum: [NEW, IN_PROGRESS, RESOLVED]
    PagedPrioritizedSignals:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PrioritizedSignal'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
    SignalResponse:
      type: object
      required:
        - id
        - title
        - description
        - category
        - status
        - priorityScore
        - scoreBreakdown
        - latitude
        - longitude
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
        priorityScore:
          type: number
        scoreBreakdown:
          $ref: '#/components/schemas/ScoreBreakdown'
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    SignalMeta:
      type: object
      required:
        - totalSignals
        - unresolvedSignals
      properties:
        totalSignals:
          type: integer
          format: int64
        unresolvedSignals:
          type: integer
          format: int64
        lastUpdatedAt:
          type: string
          format: date-time
...